implement mandatory indicator globalType
